fix: allocate VariantItems colors after annotations are loaded for each item
